requirements.txt
 
django==5.1.6
djangorestframework
django-cors-headers
transformers
pint
requests
python-dotenv